requirements.txt
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
pillow==10.3.0